feat: NovelController 'url id == message body id' 검증 로직 추가.

+++ src/main/java/com/ham/netnovel/novel/NovelController.java
@@ -8,12 +8,12 @@ import com.ham.netnovel.novel.dto.NovelUpdateDto;
 import com.ham.netnovel.novel.service.NovelService;
 import jakarta.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 @Slf4j
 @RestController
@@ -45,15 +45,26 @@ public class NovelController {
     //ToDo updateNovel, deleteNovel MessageBody와 Url에 모두 novelId를 보내는 문제 해결해야 함
     @PutMapping("/{novelId}")
     public ResponseEntity<NovelResponseDto> updateNovel(
-            @PathVariable("novelId") Long novelId,
+            @PathVariable("novelId") Long pathNovelId,
             @Valid @RequestBody NovelUpdateDto reqBody) {
+
+        //url의 novelId와 reqBody의 novelId가 같은지 검증
+        if (!pathNovelId.equals(reqBody.getNovelId())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Path Variable Id != Message Body Id");
+        }
+
         return ResponseEntity.ok(novelService.updateNovel(reqBody));
     }
 
     @DeleteMapping("/{novelId}")
     public ResponseEntity<NovelResponseDto> deleteNovel(
-            @PathVariable("novelId") Long novelId,
+            @PathVariable("novelId") Long pathNovelId,
             @Valid @RequestBody NovelDeleteDto reqBody) {
+
+        //url의 novelId와 reqBody의 novelId가 같은지 검증
+        if (!pathNovelId.equals(reqBody.getNovelId())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Path Variable Id != Message Body Id");
+        }
         return ResponseEntity.ok(novelService.deleteNovel(reqBody));
     }
 }
